build: enforce tslint
Previously, `tslint` ran during `npm run serve`, but 1) produced many unaddressed errors, and 2) was not checked in the CI pipeline. As a result, the errors are easy to miss.

This commit enforces `tslint`, by changing disagreeable configuration and fixing violating files. It also adds configuration to the CI pipeline to check lint.