feat(training): switch training cards to expansion panels
The current goal for the UI is to show *all* of the exercises available, in the order that they should be reviewed. Then if the viewer wants to interleave exercises, they can simply open the next one. If they want to skip an exercise, they can just collapse it and ignore it.

Future improvements include

* Implementing the 'settings' button.
* Showing whether the exercise has been completed in the collapsed view, probably with a checkmark icon instead of the chevron that appears by default.
* Expanding the first entry by default.
* Collapsing the current entry and expanding the next entry when recording.