feat(training): add variants of techs
Many techs have different variants for practice purposes. For example, short-hop fast-fall aerials can differ in the distance to jump and in the aerial type. We want to practice each of those variants differently. Nonetheless, they are logically more or less the same type of tech, so they should share most of the same metadata and instructions.