fix: scroll to top of page when switching routes
Currently, whenever you click the "learn tech" button from the main training page, it takes you to the learn page. But it also preserves the position on the page from before changing views.

Since you may have scrolled down a bit to view a different exercise, this meant that clicking the "learn tech" button would scroll you down the same amount on the resulting page.

Fix this by setting the `scrollBehavior` setting for Vue-router, so that it preserves the saved position for any pages in the history (from going back/forward), but otherwise pushes you to the top of the page.