feat(availability): calculate tech availability
The pedagogical idea is that you shouldn't try to learn new tech without first having a grasp of the tech it's based on. You can't learn to short-hop fast-fall aerial before you've learned how to short-hop fast-fall, and in turn before you've learned to short-hop.

We say that short-hop fast-fall aerials "depend on" short-hop fast-falls, and that short-hop fast-falls "depend on" short-hops. We should only render tech for the user to practice based on what's they've already demonstrated some mastery of. The plan is to consider tech "satisfactory" if they've had at least one practice set for that tech where they got good results.

The dependency graph is slightly complicated by the fact that tech can have variants, and some of the variants should be learned before others. For example, short-hopping in-place should probably be learned before short-hopping back and forth. We logically represent each variant of each tech as a separate node in the dependency graph, which has its own dependencies on other nodes.

The dependencies for a tech can be calculated with the `getTechDependencies` function. I tried representing the entire dependency graph as a single data structure, but that got tedious since there's so many combinations of tech and variants. Instead, it's represented programmatically.