feat(training): update hopping-related techs
This commit adds `facing`, `fall`, and `hop` as tech variants. Instead of "short-hop fast-fall aerials", we have the tech "falling aerials", which can be short- or full-hopped, and normal- or fast-falled.

Theoretically, "short-hop" and "full-hop" could be variants of the same tech, but they're so different in execution that it's worth keeping them separate.

Fast-falling could also be a variant of the short-hop and full-hop techs, but it's also so different in execution that it's worth treating as a different skill.