feat(training): store state with Vuex
For now, we store state locally in the browser. We can sync it to the cloud later.

One of the things that I didn't like about Stronglifts is that it dropped my exercise history when I signed into the cloud. The architecture of the app is designed to allow merging multiple sources of the user having exercised.

The idea is to store an append-only log of exercise entries. These will be keyed by some kind of unique identifier. When merging two sets of entries, we simply drop any entries with duplicate identifiers.

Derived data, such as data about which exercise to practice next, may have to be recomputed after a sync, or it will have to be kept up-to-date in an incremental manner.