feat(training): add spaced repetition system
Spaced repetition is an effective method for memorizing facts. The user is given a set of flashcards to review periodically. Depending on their performance, they then set the flashcard to be reviewed a short or long time from now. This way, the entire flashcard set doesn't need to be reviewed every session.

I don't know for certain if it works for training motor skills, but the app is predicated on the idea that it does.

The spaced-repetition system is based on the SuperMemo-2 algorithm.