feat(home): create home page and navigation drawer
Creates a home page with some details about the website and a call-to-action. Also creates a navigation drawer. I did this so that I could put debugging output in the "stats" page.

It was difficult to create the logo using CSS only. See the comments in `Home.vue` for more details.