feat(metadata): add character metadata
This is the listing of characters available in Super Smash Bros. Ultimate. I've also made sure to key them by game, in case we want to support Melee in the future.

There's currently a `displayName` attribute. I might also add some attributes corresponding to icon asset paths or similar, if it comes to be necessary.