build: update `browserslist` with recent browsers
There seems to be some kind of bug in the generated code where the app silently fails to dispatch Vuex (asynchronous) actions, but only in prod builds. I have no idea what the root cause is, but targeting newer browser versions which have native async/await seems to no longer have the issue. This is lamentable, but I definitely don't have the time to track down a Babel bug.