fix: sidebar links now work in all subdirectories
Writing `to="training"` redirects one to `/training` in the current directory. This happened to work for most routes since the name corresponded to the path. However, it doesn't work when you're at `/tech/<something>`, as it tries to load `/tech/training` instead.

What I had originally intended to do was to route the user to the route by its name, not its path.