style: reorder learn-exercise cards
Closes #25.

User testing indicated that the ideal ordering would be

* description
* exercise
* inputs

except that the description is typically too long to go first. The compromise is to put the exercise first and the description reasonably high up.

The "inputs" should immediately follow "exercise". So the final ordering of the above is:

* exercise
* inputs
* description

It was also agreed that the video should go last, since it's the "heaviest" component. The other cards aren't too order-sensitive.