fix: update GSP calculation for Smash Ultimate 8.1.0
Smash Ultimate patch 8.1.0 changes the cut-off for Elite Smash from the top 3.5% to the top 5%. We have to update the GSP calculation accordingly.

Previously, we calculated GSP by constructing a normal distribution and doing math to estimate the percentile. This has the advantage that it doesn't produce a greater-than-100th percentile if the user provides a larger GSP than we were expecting (but note that the user may still be correct, because the GSP calculation is based on user reports, which may be out of date). However, it also means that we consistently underestimate user percentiles compared to actual date (as the number of online users is not infinite, but a lower-bounded quantity).

This changes the calculation to just rely on elitegsp.com's calculation, which we're counting on to be correct. We could also theoretically examine the timestamp and compare it to the release date of Smash Ultimate 8.1.0 and attempt to correct it ourselves that way, but this is probably unnecessary.